+++ web-frontend/modules/builder/components/elements/components/collectionField/LinkField.vue
@@ -2,7 +2,7 @@
   <a
     class="link-button-element-button"
     v-bind="extraAttr"
-    @click="onButtonFieldClick($event)"
+    @click.prevent="onButtonFieldClick($event)"
   >
     {{ realLinkName }}
   </a>
@@ -17,30 +17,24 @@ export default {
       type: String,
       required: true,
     },
+    isExternalLink: {
+      type: Boolean,
+      required: true,
+    },
+    navigationType: {
+      type: String,
+      required: true,
+    },
     linkName: {
       type: String,
       required: true,
     },
   },
   computed: {
-    realUrl() {
-      if (!this.url) {
-        return ''
-      }
-
-      if (this.url.startsWith('/') && this.mode === 'preview') {
-        return `/builder/${this.builder.id}/preview${this.url}`
-      } else {
-        return this.url
-      }
-    },
-    isExternalLink() {
-      return !this.realUrl.startsWith('/')
-    },
     extraAttr() {
       const attr = {}
-      if (this.realUrl) {
-        attr.href = this.realUrl
+      if (this.url) {
+        attr.href = this.url
       }
       if (this.isExternalLink) {
         attr.rel = 'noopener noreferrer'
@@ -57,16 +51,13 @@ export default {
   },
   methods: {
     onButtonFieldClick(event) {
-      if (this.mode === 'editing') {
-        event.preventDefault()
+      if (this.mode === 'editing' || !this.url) {
         return
       }
-
-      if (!this.url) {
-        event.preventDefault()
-      } else if (!this.isExternalLink) {
-        event.preventDefault()
-        this.$router.push(this.realUrl)
+      if (this.navigationType === 'custom') {
+        window.location.href = this.url
+      } else if (this.navigationType === 'page') {
+        this.$router.push(this.url)
       }
     },
   },