+++ web-frontend/modules/builder/elementTypes.js
@@ -12,16 +12,12 @@ import InputTextElementForm from '@baserow/modules/builder/components/elements/c
 import TableElement from '@baserow/modules/builder/components/elements/components/TableElement.vue'
 import TableElementForm from '@baserow/modules/builder/components/elements/components/forms/general/TableElementForm.vue'
 
-import {
-  ELEMENT_EVENTS,
-  PAGE_PARAM_TYPE_VALIDATION_FUNCTIONS,
-} from '@baserow/modules/builder/enums'
+import { ELEMENT_EVENTS } from '@baserow/modules/builder/enums'
 import { ensureBoolean } from '@baserow/modules/core/utils/validator'
 import ColumnElement from '@baserow/modules/builder/components/elements/components/ColumnElement'
 import ColumnElementForm from '@baserow/modules/builder/components/elements/components/forms/general/ColumnElementForm'
 import _ from 'lodash'
 import DefaultStyleForm from '@baserow/modules/builder/components/elements/components/forms/style/DefaultStyleForm'
-import { compile } from 'path-to-regexp'
 import ButtonElement from '@baserow/modules/builder/components/elements/components/ButtonElement'
 import ButtonElementForm from '@baserow/modules/builder/components/elements/components/forms/general/ButtonElementForm'
 import { ClickEvent, SubmitEvent } from '@baserow/modules/builder/eventTypes'
@@ -101,13 +97,27 @@ export class ElementType extends Registerable {
   }
 
   /**
-   * This hook allows you to change the values given by the form of the element before
-   * they are sent to the backend to update the element.
-   *
-   * @param {object} values - The values of the element
-   * @returns {*}
+   * Allow to hook into default values for this element type at element creation.
+   * @param {object} values the current values for the element to create.
+   * @returns an object containing values updated with the default values.
    */
-  prepareValuesForRequest(values) {
+  getDefaultValues(page, values) {
+    // By default if an element is inside a container we apply the
+    // `.getDefaultChildValues()` method of the parent to it.
+    if (values?.parent_element_id) {
+      const parentElement = this.app.store.getters['element/getElementById'](
+        page,
+        values.parent_element_id
+      )
+      const parentElementType = this.app.$registry.get(
+        'element',
+        parentElement.type
+      )
+      return {
+        ...values,
+        ...parentElementType.getDefaultChildValues(page, values),
+      }
+    }
     return values
   }
 
@@ -187,6 +197,17 @@ export class ContainerElementType extends ElementType {
 
   get defaultPlaceInContainer() {
     throw new Error('Not implemented')
+  }
+
+  /**
+   * Returns the default value when creating a child element to this container.
+   * @param {Object} page The current page object
+   * @param {Object} values The values of the to be created element
+   * @returns the default values for this element.
+   */
+  getDefaultChildValues(page, values) {
+    // By default an element inside a container should have no left and right padding
+    return { style_padding_left: 0, style_padding_right: 0 }
   }
 }
 
@@ -438,39 +459,6 @@ export class LinkElementType extends ElementType {
     } //
 
     return false
-  }
-
-  static getUrlFromElement(element, builder, resolveFormula) {
-    if (element.navigation_type === 'page') {
-      if (!isNaN(element.navigate_to_page_id)) {
-        const page = builder.pages.find(
-          ({ id }) => id === element.navigate_to_page_id
-        )
-
-        // The builder page list might be empty or the page has been deleted
-        if (!page) {
-          return ''
-        }
-
-        const paramTypeMap = Object.fromEntries(
-          page.path_params.map(({ name, type }) => [name, type])
-        )
-
-        const toPath = compile(page.path, { encode: encodeURIComponent })
-        const pageParams = Object.fromEntries(
-          element.page_parameters.map(({ name, value }) => [
-            name,
-            PAGE_PARAM_TYPE_VALIDATION_FUNCTIONS[paramTypeMap[name]](
-              resolveFormula(value)
-            ),
-          ])
-        )
-        return toPath(pageParams)
-      }
-    } else {
-      return resolveFormula(element.navigate_to_url)
-    }
-    return ''
   }
 }